elev8: Connect to EVAS_CALLBACK_FREE event.
This callback is called after the Evas_Object has been freed from the EFL
side.  When this event is received, a debug message is printed and the
pointer to said Evas_Object is nullified.

Hopefully magic checks inside Elementary will print the expected object type
when it finds a null pointer, giving a hint of where to look for possible
errors.

This shouldn't happen in normal circumstances, but it is a good measure to
avoid crashes and hunt for bugs.

SVN revision: 72620