elev8: Add CElmThumb thumbnail widget
this patch adds basic support for the elm_thumb widget to elev8. A demo
script can be found in data/javascript/.

Not yet implemented are the ethumb_client_* calls. I'm not sure if that
should be its own class or if the functions should be included in here.

Signed-off-by: Daniel Willmann <d.willmann@samsung.com>

SVN revision: 72622