elev8: Server mode.
Summary:
Now it's possible to create an "elev8 server" that will preload all
modules and, when one requests an app, the server will fork itself to
run the app, in order to avoid loading all modules each time an app is
executed.
 - `elev8 --server` shall create the daemon server.
 - `elev8 --shutdown` will kill this server
 - `elev8 file.js` will run the file.
When runnig an app (last option), if a server is found, elev8  will close
and the server will spawn a process to run the file. If it can't find a
server, it will run the file "locally", pretty much like the old way.
Finally, the debug option will just run "locally".

 Author:    ederson <ederson@profusion.mobi>

SVN revision: 79900